+++ k-PPD-ERT/party_class.py
@@ -103,7 +103,7 @@ class Party:
         #  percentage of attributes randomly selected
         att_perc = self.att_perc
         #  number of attributes randomly selected
-        num_rnd_att = round(att_perc * num_att)
+        num_rnd_att = np.int(round(att_perc * num_att))
         # indices of randomly selected attributes
         rand_attribute_index = [random.randint(0, num_att - 1) for p in range(0, num_rnd_att)]
         # SELECT RANDOM SPLITS
